requirements.txt
fastapi[standard]
uvicorn[standard]
requests
sqlalchemy
python-dotenv
